+++ ZbazinersBar/Models/ViewModels/PagingInfo.cs
@@ -7,6 +7,6 @@ namespace ZbazinersBar.Models.ViewModels {
         public int CurrentPage { get; set; }
 
         public int TotalPages =>
-            (int)Math.Ceiling((decimal)TotalItems / ItemsPerPage);
+            (int)Math.Ceiling((float)TotalItems / ItemsPerPage);
     }
 }
